backend/requirements.txt
Babel==2.7.0
beautifulsoup4~=4.8.0
bleach~=3.1.0
bleach-whitelist~=0.0.10
celery~=4.3.0
Django==2.2.4
django-bootstrap4~=0.0.8
django-cookie-law~=1.0.1
django-cors-headers~=3.1.0
django-money==0.15
django-phonenumber-field==3.0.1
django-versatileimagefield~=1.10
django-rest-swagger~=2.2.0
djangorestframework~=3.10.3
drf-yasg~=1.16.1
gunicorn~=19.9.0
django-extra-fields~=1.2.4
django-rest-knox~=4.1.0
html2text~=2018.1.9
ipaddress~=1.0.22
django-markup~=1.4.1
markdown~=3.1.1
packaging~=19.1
phonenumbers==8.10.14
pyyaml~=5.1.1
psycopg2-binary~=2.8.3
pytz==2019.1
requests~=2.22.0
setuptools>=41.0.1
sendgrid-django~=4.2.0
sentry-sdk==0.10.2
sqlparse==0.3.0
